Thesis/Step_1_Voice/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydub==0.25.1
ffmpeg-python==0.2.0
TTS==0.21.2
numpy==1.22.0
numba==0.60.0
torch==2.5.1
torchaudio==2.5.1
google-generativeai==0.3.2
python-multipart==0.0.9
python-dotenv==1.0.1
sounddevice==0.4.6
soundfile==0.12.1